APPSERV-151 Add command to generate CSR and some minor refactors
