+++ nucleus/admin/server-mgmt/src/main/java/com/sun/enterprise/admin/servermgmt/cli/LocalServerCommand.java
@@ -64,6 +64,8 @@ import com.sun.enterprise.universal.process.ProcessUtils;
 import com.sun.enterprise.universal.xml.MiniXmlParser;
 import com.sun.enterprise.universal.xml.MiniXmlParserException;
 import com.sun.enterprise.util.HostAndPort;
+import com.sun.enterprise.util.StringUtils;
+import com.sun.enterprise.util.SystemPropertyConstants;
 import com.sun.enterprise.util.io.FileUtils;
 import com.sun.enterprise.util.io.ServerDirs;
 
@@ -639,5 +641,25 @@ public abstract class LocalServerCommand extends CLICommand {
         logger.warning("Could not determine if data grid encryption is enabled - " +
                 "you will need to regenerate the encryption key if it is");
         return false;
+    }
+
+    /**
+     * Gets the GlassFish installation root (using property com.sun.aas.installRoot),
+     * first from asenv.conf.  If that's not available, then from java.lang.System.
+     *
+     * @return path of GlassFish install root
+     * @throws CommandException if the GlassFish install root is not found
+     */
+    protected String getInstallRootPath() throws CommandException {
+        String installRootPath = getSystemProperty(SystemPropertyConstants.INSTALL_ROOT_PROPERTY);
+
+        if (!StringUtils.ok(installRootPath)) {
+            installRootPath = System.getProperty(SystemPropertyConstants.INSTALL_ROOT_PROPERTY);
+        }
+
+        if (!StringUtils.ok(installRootPath)) {
+            throw new CommandException("noInstallDirPath");
+        }
+        return installRootPath;
     }
 }
